Tests for Binary Search.

+++ src/main/java/pl/arturszejna/algorithms/numeralSequences/sumOfNumbersInRange/SumOfNumbersInRange.java
@@ -10,7 +10,7 @@ public class SumOfNumbersInRange {
 
     //złozoność obliczeniowa - O(n)
     //złozoność pamięciowa - O(1)
-    public static Double sumOfRangeWithLoop(int start, int end){
+    public static double sumOfRangeWithLoop(int start, int end){
         long sum = 0L;
         for (int i = start; i <= end; i++){
             sum = sum + i;
@@ -20,9 +20,9 @@ public class SumOfNumbersInRange {
 
     //złozoność obliczeniowa - O(1)
     //złozoność pamięciowa - O(1)
-    public static Double sumOfRange(int start, int end){
-        Double startDouble = (double) start;
-        Double endDouble = (double) end;
-        return (((startDouble + endDouble) / 2) * (endDouble - startDouble + 1));
+    public static double sumOfRange(int start, int end){
+        double startDouble = start;
+        double endDouble = end;
+        return ((startDouble + endDouble) / 2) * (endDouble - startDouble + 1);
     }
 }
